scraper/requirements.txt
scrapy==2.11.2
celery==5.4.0
redis==5.0.8
requests==2.32.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
